upgrade to using lire1.0b2

+++ src/main/java/org/elasticsearch/index/query/image/AbstractImageScorer.java
@@ -1,6 +1,6 @@
 package org.elasticsearch.index.query.image;
 
-import net.semanticmetadata.lire.imageanalysis.LireFeature;
+import net.semanticmetadata.lire.imageanalysis.features.LireFeature;
 import org.apache.lucene.index.BinaryDocValues;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.LeafReader;
@@ -47,14 +47,14 @@ public abstract class AbstractImageScorer extends Scorer {
             LireFeature docFeature = lireFeature.getClass().newInstance();
             docFeature.setByteArrayRepresentation(bytesRef.bytes);
 
-            float distance = lireFeature.getDistance(docFeature);
-            float score;
-            if (Float.compare(distance, 1.0f) <= 0) { // distance less than 1, consider as same image
+            double distance = lireFeature.getDistance(docFeature);
+            double score;
+            if (Double.compare(distance, 1.0f) <= 0) { // distance less than 1, consider as same image
                 score = 2f - distance;
             } else {
                 score = 1 / distance;
             }
-            return score * boost;
+            return (float) score * boost;
         } catch (Exception e) {
             throw new ElasticsearchImageProcessException("Failed to calculate score", e);
         }
